refactor(cli): move lab port-access helpers into lab_render
Resolves the SonarCloud quality-gate failures on the dev->main promotion
(#739). All six smells were in src/aptl/cli/lab.py:

- File length 563 > 500 lines: the host-port access-summary helpers move
  out to the existing cli/lab_render.py rendering module. lab.py is now
  399 lines.
- `_live_resolved_ports` cognitive complexity 34 > 15 and cyclomatic
  14 > 10: split into `_cli_backend`, `_container_resolved_ports`, and
  `_binding_to_resolved_port`, each shallowly nested. The public
  `live_resolved_ports` is now a thin loop.
- Three `# noqa: BLE001` blind-except suppressions dropped: ruff only
  enforces C901 (complexity), so the bare `except Exception` — required
  so a bad/missing project config never aborts `aptl lab info` — is
  lint-clean without the suppression comments.

Behaviour is unchanged: `_emit_lab_access_summary` -> `emit_lab_access_summary`
and `_live_resolved_ports` -> `live_resolved_ports` are the same logic, now
imported by lab.py's `start`/`info` commands. Tests updated to the new
import paths; no assertion changes.
